authors: consolidate workflows into one
* Moves actions and workflows from author module to a common
  "Author" workflow in the workflows module.
