+++ inspirehep/modules/workflows/actions/__init__.py
@@ -25,6 +25,7 @@
 
 from .hep_approval import HEPApproval
 from .merge_approval import MergeApproval
+from .author_approval import AuthorApproval
 
 
-__all__ = ('HEPApproval', 'MergeApproval')
+__all__ = ('HEPApproval', 'MergeApproval', 'AuthorApproval')
